HBX-1836: Move the ant tasks from the core hibernate tools into the hibernate-tools-ant project - Move the old ant tasks from 'org.hibernate.tool.ant.old' back to 'org.hibernate.tool.ant' for now
