+++ ant/src/main/java/org/hibernate/tool/ant/Hbm2CfgXmlExporterTask.java
@@ -2,7 +2,7 @@
  * Created on 25-Feb-2005
  *
  */
-package org.hibernate.tool.ant.old;
+package org.hibernate.tool.ant;
 
 import org.hibernate.tool.api.export.Exporter;
 import org.hibernate.tool.internal.export.cfg.CfgExporter;
