+++ ant/src/main/java/org/hibernate/tool/ant/HbmLintExporterTask.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.ant.old;
+package org.hibernate.tool.ant;
 
 import org.hibernate.tool.api.export.Exporter;
 import org.hibernate.tool.internal.export.lint.HbmLintExporter;
