+++ ant/src/main/java/org/hibernate/tool/ant/JDBCConfigurationTask.java
@@ -2,7 +2,7 @@
  * Created on 15-Feb-2005
  *
  */
-package org.hibernate.tool.ant.old;
+package org.hibernate.tool.ant;
 
 import java.io.File;
 import java.lang.reflect.Constructor;
