+++ ant/src/main/java/org/hibernate/tool/ant/JavaFormatterTask.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.ant.old;
+package org.hibernate.tool.ant;
 
 import java.io.BufferedInputStream;
 import java.io.File;
